Separate whitelisting from deployment
